refactor: 구간 인수테스트를 성공테스트와 실패테스트로 분리
