+++ src/test/java/nextstep/subway/acceptance/SectionAcceptanceSuccessTest.java
@@ -7,16 +7,17 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 
-import java.util.HashMap;
 import java.util.Map;
 
 import static nextstep.subway.acceptance.LineSteps.*;
+import static nextstep.subway.acceptance.SectionSteps.createLineCreateParams;
+import static nextstep.subway.acceptance.SectionSteps.createSectionCreateParams;
 import static nextstep.subway.acceptance.StationSteps.지하철역_생성_요청;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("지하철 구간 관리 기능")
-class SectionAcceptanceTest extends AcceptanceTest {
+class SectionAcceptanceSuccessTest extends AcceptanceTest {
     private Long 신분당선;
 
     private Long 강남역;
@@ -73,44 +74,6 @@ class SectionAcceptanceTest extends AcceptanceTest {
     }
 
     /**
-     * When 새로운 구간을 중간에 삽입할 때, 기존 구간의 길이보다 크거나 같으면
-     * Then 노선에 새로운 구간이 추가되지 않는다.
-     */
-    @DisplayName("새로운 구간 삽입 시 기존 구간의 길이보다 크거나 같으면 새로운 구간이 추가되지 않는다.")
-    @Test
-    void addLineSectionWithUpStationFailOnLongerDistance() {
-        // when
-        Long 양재역 = 지하철역_생성_요청("양재역").jsonPath().getLong("id");
-        ExtractableResponse<Response> 구간의_길이가_클_때_응답 = 지하철_노선에_지하철_구간_생성_요청(신분당선, createSectionCreateParams(강남역, 양재역, 15));
-        ExtractableResponse<Response> 구간의_길이가_같을_때_응답 = 지하철_노선에_지하철_구간_생성_요청(신분당선, createSectionCreateParams(강남역, 양재역, 10));
-
-        // then
-        assertThat(구간의_길이가_클_때_응답.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-        assertThat(구간의_길이가_같을_때_응답.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    /**
-     * Given 노선에 구간을 등록한다.
-     * When 새로운 구간을 추가 할 때, 기존 구간에 상행역과 하행역이 모두 포함되면
-     * Then 노선에 구간이 추가되지 않는다.
-     */
-    @DisplayName("새로운 구간 등록 시, 먼저 상행역과 하행역이 구간으로 이미 등록되어 있으면 실패한다.")
-    @Test
-    void addLineSectionFailWithAlreadyAdded() {
-        // given
-        Long 양재역 = 지하철역_생성_요청("양재역").jsonPath().getLong("id");
-        Long 신사역 = 지하철역_생성_요청("신사역").jsonPath().getLong("id");
-        지하철_노선에_지하철_구간_생성_요청(신분당선, createSectionCreateParams(강남역, 양재역, 3));
-        지하철_노선에_지하철_구간_생성_요청(신분당선, createSectionCreateParams(신사역, 강남역, 3));
-
-        // when
-        ExtractableResponse<Response> 구간_생성_요청_응답 = 지하철_노선에_지하철_구간_생성_요청(신분당선, createSectionCreateParams(신사역, 양재역, 3));
-
-        // then
-        assertThat(구간_생성_요청_응답.statusCode()).isEqualTo(HttpStatus.BAD_REQUEST.value());
-    }
-
-    /**
      * When 기존 구간의 upStation과 일치하는 downStation을 가진 구간이 추가 요청이 되면
      * Then 노선에 새로운 구간이 추가된다
      */
@@ -148,24 +111,5 @@ class SectionAcceptanceTest extends AcceptanceTest {
         ExtractableResponse<Response> response = 지하철_노선_조회_요청(신분당선);
         assertThat(response.statusCode()).isEqualTo(HttpStatus.OK.value());
         assertThat(response.jsonPath().getList("stations.id", Long.class)).containsExactly(강남역, 판교역);
-    }
-
-    private Map<String, String> createLineCreateParams(Long upStationId, Long downStationId) {
-        Map<String, String> lineCreateParams;
-        lineCreateParams = new HashMap<>();
-        lineCreateParams.put("name", "신분당선");
-        lineCreateParams.put("color", "bg-red-600");
-        lineCreateParams.put("upStationId", upStationId + "");
-        lineCreateParams.put("downStationId", downStationId + "");
-        lineCreateParams.put("distance", 10 + "");
-        return lineCreateParams;
-    }
-
-    private Map<String, String> createSectionCreateParams(Long upStationId, Long downStationId, Integer distance) {
-        Map<String, String> params = new HashMap<>();
-        params.put("upStationId", upStationId + "");
-        params.put("downStationId", downStationId + "");
-        params.put("distance", distance + "");
-        return params;
     }
 }
